Server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
easyocr==1.7.1
en-core-med7-lg @ https://huggingface.co/kormilitzin/en_core_med7_lg/resolve/main/en_core_med7_lg-any-py3-none-any.whl#sha256=2c8ce371b7ee8b3c42cce7c05031ab165e6e88cc3ed136797eb3c492af96001f
fastapi==0.110.0
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
lazy_loader==0.3
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==23.2
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.2.0
preshed==3.0.9
pyclipper==1.3.0.post5
pydantic==1.10.14
pydantic_core==2.16.3
python-bidi==0.4.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
shapely==2.0.3
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.36.3
sympy==1.12
textblob==0.18.0.post0
thinc==8.1.12
threadpoolctl==3.3.0
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typer==0.7.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
wasabi==0.10.1
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
paddleocr>=2.7.1
paddlepaddle==2.6.0
